fix: default strategy screen not loading
