+++ frontend/src/component/project/Project/ProjectSettings/ProjectDefaultStrategySettings/ProjectEnvironment/ProjectEnvironmentDefaultStrategy/ProjectEnvironmentDefaultStrategy.tsx
@@ -2,13 +2,11 @@ import { useRequiredPathParam } from 'hooks/useRequiredPathParam';
 import { StrategyItemContainer } from 'component/common/StrategyItemContainer/StrategyItemContainer';
 import PermissionIconButton from 'component/common/PermissionIconButton/PermissionIconButton';
 import { UPDATE_FEATURE_STRATEGY } from 'component/providers/AccessProvider/permissions';
-import { Link, Route, Routes, useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { Edit } from '@mui/icons-material';
 import { StrategyExecution } from 'component/feature/FeatureView/FeatureOverview/FeatureOverviewEnvironments/FeatureOverviewEnvironment/EnvironmentAccordionBody/StrategyDraggableItem/StrategyItem/StrategyExecution/StrategyExecution';
 import { ProjectEnvironmentType } from 'interfaces/environments';
 import React, { useMemo } from 'react';
-import EditDefaultStrategy from './EditDefaultStrategy';
-import { SidebarModal } from 'component/common/SidebarModal/SidebarModal';
 import { CreateFeatureStrategySchema } from 'openapi';
 
 interface ProjectEnvironmentDefaultStrategyProps {
@@ -41,7 +39,6 @@ const ProjectEnvironmentDefaultStrategy = ({
     environment,
     description,
 }: ProjectEnvironmentDefaultStrategyProps) => {
-    const navigate = useNavigate();
     const projectId = useRequiredPathParam('projectId');
     const { environment: environmentId, defaultStrategy } = environment;
 
@@ -50,13 +47,9 @@ const ProjectEnvironmentDefaultStrategy = ({
         environmentId
     );
 
-    const path = `/projects/${projectId}/settings/default-strategy`;
-
     const strategy: CreateFeatureStrategySchema = useMemo(() => {
         return defaultStrategy ? defaultStrategy : DEFAULT_STRATEGY;
     }, [JSON.stringify(defaultStrategy)]);
-
-    const onSidebarClose = () => navigate(path);
 
     return (
         <>
@@ -83,20 +76,6 @@ const ProjectEnvironmentDefaultStrategy = ({
             >
                 <StrategyExecution strategy={strategy} />
             </StrategyItemContainer>
-            <Routes>
-                <Route
-                    path="edit"
-                    element={
-                        <SidebarModal
-                            label="Edit feature strategy"
-                            onClose={onSidebarClose}
-                            open
-                        >
-                            <EditDefaultStrategy />
-                        </SidebarModal>
-                    }
-                />
-            </Routes>
         </>
     );
 };
